fix: (home) 읽고 있는 책 조회 API 변경사항 반영
- 서비스 요구사항에 따라 readingPercent 값을 int형으로 보내도록 변경

+++ server/src/main/java/com/codecozy/server/dto/response/GetReadingResponse.java
@@ -6,7 +6,7 @@ public record GetReadingResponse(
         String cover,
         String title,
         String author,
-        Double readingPercent,
+        int readingPercent,
         int totalPage,
         int readPage,
         Boolean isHidden,
